refactor: 프로퍼티 클래스를 올바른 패키지로 이동
