+++ src/main/java/com/b4f2/pting/config/properties/KakaoOAuthProperties.java
@@ -1,4 +1,4 @@
-package com.b4f2.pting.domain.properties;
+package com.b4f2.pting.config.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
